Added restriction page ability based on the login

+++ studyapp/views.py
@@ -6,6 +6,9 @@ from django.db.models import Q
 from django.contrib.auth import authenticate,login,logout
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+
 # Create your views here.
 
 # rooms=[
@@ -16,6 +19,10 @@ from django.contrib import messages
 
 #####I created this view in the last after DeleteRoom
 def loginPage(request):
+   page='login'
+   if request.user.is_authenticated:
+      return redirect('home')
+   
    if request.method=='POST':
       username=request.POST.get('username')
       password=request.POST.get('password')
@@ -32,12 +39,19 @@ def loginPage(request):
          return redirect('home')
       else:
          messages.error("Invalid credentials")
-   context={}
+   context={'page':page}
    return render(request,'studyapp/login_register.html',context)
 
 def logoutUser(request):
    logout(request)
    return redirect('home')
+
+def registerUser(request):
+   page='register'
+
+   context={'page':page}
+   return render(request,'studyapp/login_register.html',context)
+
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q')!=None else ''
@@ -56,6 +70,7 @@ def room(request,pk):
   context={'room':room}
   return render(request,'studyapp/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
    form=RoomForm()
    if request.method=='POST':
@@ -68,10 +83,13 @@ def createRoom(request):
    context={'form':form}
    return render(request,'studyapp/room_form.html',context)
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
    room=Room.objects.get(id=pk)
    form=RoomForm(instance=room)
+
+   if request.user!=room.host:
+      return HttpResponse("Your'e not allowed to do this!!!")
 
    if request.method=='POST':
       form=RoomForm(request.POST,instance=room)
@@ -81,8 +99,13 @@ def updateRoom(request,pk):
    context={'form':form}
    return render(request,'studyapp/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
    room=Room.objects.get(id=pk)
+
+   if request.user!=room.host:
+      return HttpResponse("Your'e not allowed to do this!!!")
+   
    if request.method=='POST':
       room.delete()
       return redirect('home')
